+++ CHANGELOG.md
@@ -4,14 +4,39 @@ X.Y.Z Release notes
 * None
 
 ### Enhancements
+* Improve performance of the RPC worker for chrome debugging.
+* Added `Realm.deleteFile` for deleting a Realm (#363).
+* `Realm.Sync.User.login`, `Realm.Sync.User.register`, and `Realm.Sync.User.registerWithProvider` both support _both_ callbacks and Promise return values
+
+### Bug fixes
+* None
+
+1.11.1 Release notes (2017-9-1)
+=============================================================
+### Breaking changes
+* None
+
+### Enhancements
+* None
+
+### Bug fixes
+* Fix accessToken.
+
+1.11.0 Release notes (2017-8-31)
+=============================================================
+### Breaking changes
+* None
+
+### Enhancements
 * Added methods `Realm.beginTransaction()`, `Realm.commitTransaction()`, `Realm.cancelTransaction()` to manually control write transactions.
 * Added property `Realm.isInTransaction` which indicates if write transaction is in progress.
 * Added `shouldCompactOnLaunch` to configuration (#507).
 * Added `Realm.compact()` for manually compacting Realm files.
-* `Realm.Sync.User.login`, `Realm.Sync.User.register`, and `Realm.Sync.User.registerWithProvider` both support _both_ callbacks and Promise return values
+* Added various methods for permission management (#1204).
 
 ### Bug fixes
 * Removed `loginWithProvider` from TypeScript definition files. This API never existed and was incorrectly added.
+
 
 1.10.3 Release notes (2017-8-16)
 =============================================================